jorge
Broke the wtoryline into a method we can call as the sotry progresses.
Added derived planets classes to the planets class
adjusted the weapons classes to the option bran had in the storyline
Worked on the markup. i'll show it to yall tomorrow
Came up with methods fpr the actions class. Lmk what yall think